delete post & comment in ref

+++ controllers/post.controller.js
@@ -4,10 +4,7 @@ const User = require("../models/user.model");
 
 // only logged user can get all posts
 const getAllPosts = async (req, res, next) => {
-  const posts = await Post.find()
-    .populate("user")
-    .populate("comments")
-    .populate("reviews");
+  const posts = await Post.find().populate("comments").populate("reviews");
 
   if (!posts) {
     return next(new AppError("No posts found for this user!", 404));
@@ -28,9 +25,11 @@ const getAllPosts = async (req, res, next) => {
 
 const getSinglePost = async (req, res, next) => {
   const { id } = req.params;
-  const userId = req.user.id;
+  // const userId = req.user.id;
   try {
-    const singlePost = await Post.find({ _id: id });
+    const singlePost = await Post.find({ _id: id })
+      .populate("comments")
+      .populate("reviews");
     if (!singlePost || singlePost.length == 0) {
       return next(new AppError("Post not found", 404));
     }
@@ -89,21 +88,37 @@ const deletePost = async (req, res, next) => {
   const { id } = req.params;
   const user = req.user;
   try {
+    const post = await Post.find({ _id: id });
+    if (post.length == 0) {
+      return next(new AppError("Post not found", 404));
+    }
     if (user.role == "admin") {
       const deletedPost = await Post.findByIdAndDelete(id);
+      const updatedUser = await User.findOneAndUpdate(
+        { posts: { _id: id } },
+        { $pull: { posts: id } },
+        { new: true }
+      );
+
       if (!deletedPost) {
         return next(new AppError("Post not found", 404));
       }
-      res.status(200).json({ deletedPost });
+      res.status(200).json({ deletedPost, updatedUser });
     } else {
       const deletedPost = await Post.findOneAndDelete({
         _id: id,
         user: user.id,
       });
+      const updatedUser = await User.findOneAndUpdate(
+        { _id: user.id },
+        { $pull: { posts: id } },
+        { new: true }
+      );
+
       if (!deletedPost) {
         return next(new AppError("invalid token", 401));
       }
-      res.status(200).json({ deletedPost });
+      res.status(200).json({ deletedPost, updatedUser });
     }
   } catch (err) {
     return next(new AppError("Something went wrong!", 404));
